requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
idna==2.9
oauthlib==3.1.0
phonenumberslite==8.12.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
